paper-parser-mcp/requirements.txt
mcp>=1.0.0
marker-pdf>=0.2.6
pymupdf4llm>=0.2.9
pydantic>=2.0
Pillow>=10.0
